# Bug 575787 - [win32] Optimize Tree.setItemCount