--- a/bundles/org.eclipse.swt/Eclipse SWT/win32/org/eclipse/swt/widgets/Tree.java
+++ b/bundles/org.eclipse.swt/Eclipse SWT/win32/org/eclipse/swt/widgets/Tree.java
@@ -2069,9 +2069,8 @@ void createItem (TreeItem item, long hParent, long hInsertAfter, long hItem) {
 				shrink = true;
 				length = Math.max (4, items.length * 3 / 2);
 			}
-			TreeItem [] newItems = new TreeItem [length];
-			System.arraycopy (items, 0, newItems, 0, items.length);
-			items = newItems;
+
+			itemsGrowArray (length);
 		}
 		lastID = id + 1;
 	}
@@ -3855,6 +3854,22 @@ boolean isUseWsBorder () {
 	return true;
 }
 
+int itemsGetFreeCapacity() {
+	int count = 0;
+	for (TreeItem item : items) {
+		if (item == null)
+			count++;
+	}
+
+	return count;
+}
+
+void itemsGrowArray (int newCapacity) {
+	TreeItem [] newItems = new TreeItem [newCapacity];
+	System.arraycopy (items, 0, newItems, 0, items.length);
+	items = newItems;
+}
+
 void redrawSelection () {
 	if ((style & SWT.SINGLE) != 0) {
 		long hItem = OS.SendMessage (handle, OS.TVM_GETNEXTITEM, OS.TVGN_CARET, 0);
@@ -4137,69 +4152,125 @@ public void setItemCount (int count) {
 }
 
 void setItemCount (int count, long hParent) {
-	long hItem = OS.SendMessage (handle, OS.TVM_GETNEXTITEM, OS.TVGN_CHILD, hParent);
+	// Investigate existing items and decide what to do
+	long itemFirstChild = OS.SendMessage (handle, OS.TVM_GETNEXTITEM, OS.TVGN_CHILD, hParent);
+	long itemInsertAfter = 0;
+	int  indexInsertAfter = 0;
+	int  numInserted = 0;
+	long itemDeleteFrom = 0;
+	{
+		// Iterate to position #count and find prev/next items at this position
+		int itemCount = 0;
+		long itemPrev = OS.TVI_FIRST;
+		long itemNext = itemFirstChild;
+		while (itemNext != 0 && itemCount < count)
+		{
+			itemPrev = itemNext;
+			itemNext = OS.SendMessage (handle, OS.TVM_GETNEXTITEM, OS.TVGN_NEXT, itemNext);
+			itemCount++;
+		}
+
+		if ((itemCount == count) && (itemNext == 0)) {
+			// Exactly 'count' items, no need to do anything.
+			return;
+		} else if (itemCount == count) {
+			// Too many items, going to delete some
+			itemDeleteFrom = itemNext;
+		} else if (itemNext == 0) {
+			// Counted all items, and there is not enough, going to insert some.
+			itemInsertAfter = itemPrev;
+			indexInsertAfter = itemCount - 1;
+			numInserted = count - itemCount;
+		}
+	}
 
 	boolean redraw = false;
 	if (OS.SendMessage (handle, OS.TVM_GETCOUNT, 0, 0) == 0) {
 		redraw = getDrawing () && OS.IsWindowVisible (handle);
 		if (redraw) OS.DefWindowProc (handle, OS.WM_SETREDRAW, 0, 0);
 	}
-	int itemCount = 0;
-	while (hItem != 0 && itemCount < count) {
-		hItem = OS.SendMessage (handle, OS.TVM_GETNEXTITEM, OS.TVGN_NEXT, hItem);
-		itemCount++;
-	}
+
 	boolean expanded = false;
 	TVITEM tvItem = new TVITEM ();
 	tvItem.mask = OS.TVIF_HANDLE | OS.TVIF_PARAM;
-	if (!redraw && (style & SWT.VIRTUAL) != 0 && (hParent != OS.TVI_ROOT)) {
-		/*
-		* Bug in Windows.  Despite the fact that TVM_GETITEMSTATE claims
-		* to return only the bits specified by the stateMask, when called
-		* with TVIS_EXPANDED, the entire state is returned.  The fix is
-		* to explicitly check for the TVIS_EXPANDED bit.
-		*/
-		int state = (int)OS.SendMessage (handle, OS.TVM_GETITEMSTATE, hParent, OS.TVIS_EXPANDED);
-		expanded = (state & OS.TVIS_EXPANDED) != 0;
-	}
-	while (hItem != 0) {
-		tvItem.hItem = hItem;
-		OS.SendMessage (handle, OS.TVM_GETITEM, 0, tvItem);
-		hItem = OS.SendMessage (handle, OS.TVM_GETNEXTITEM, OS.TVGN_NEXT, hItem);
-		TreeItem item = tvItem.lParam != -1 ? items [(int)tvItem.lParam] : null;
-		if (item != null && !item.isDisposed ()) {
-			item.dispose ();
+	if (!redraw && (style & SWT.VIRTUAL) != 0) {
+		if (hParent == OS.TVI_ROOT) {
+			// Trying to call TVM_GETITEMSTATE with TVI_ROOT causes a crash.
+			// Assume that root item is always expanded.
+			expanded = true;
 		} else {
-			releaseItem (tvItem.hItem, tvItem, false);
-			destroyItem (null, tvItem.hItem);
+			/*
+			 * Bug in Windows.  Despite the fact that TVM_GETITEMSTATE claims
+			 * to return only the bits specified by the stateMask, when called
+			 * with TVIS_EXPANDED, the entire state is returned.  The fix is
+			 * to explicitly check for the TVIS_EXPANDED bit.
+			 */
+			int state = (int)OS.SendMessage (handle, OS.TVM_GETITEMSTATE, hParent, OS.TVIS_EXPANDED);
+			expanded = (state & OS.TVIS_EXPANDED) != 0;
 		}
 	}
-	if ((style & SWT.VIRTUAL) != 0) {
-		for (int i=itemCount; i<count; i++) {
-			/*
-			 * Bug 206806: Windows sends 'TVN_GETDISPINFO' when item is
-			 * being inserted. This causes 'SWT.SetData' to be sent to
-			 * user code, but user code will likely be confused by
-			 * inconsistent Tree state (because we're still inserting):
-			 * - 'getItemCount()' will be wrong
-			 * - 'Event.index' will be wrong
-			 * The workaround is to temporarily suppress 'SWT.SetData'. Note
-			 * that the boolean flag is misleadingly used for multiple
-			 * purposes. What really happens is that 'TVN_GETDISPINFO' will
-			 * queue a repaint for item and early return.
-			 */
-			if (expanded) ignoreShrink = true;
-			createItem (null, hParent, OS.TVI_LAST, 0);
-			if (expanded) ignoreShrink = false;
+
+	if (itemDeleteFrom != 0) {
+		while (itemDeleteFrom != 0) {
+			tvItem.hItem = itemDeleteFrom;
+			OS.SendMessage (handle, OS.TVM_GETITEM, 0, tvItem);
+			itemDeleteFrom = OS.SendMessage (handle, OS.TVM_GETNEXTITEM, OS.TVGN_NEXT, itemDeleteFrom);
+			TreeItem item = tvItem.lParam != -1 ? items [(int)tvItem.lParam] : null;
+			if (item != null && !item.isDisposed ()) {
+				item.dispose ();
+			} else {
+				releaseItem (tvItem.hItem, tvItem, false);
+				destroyItem (null, tvItem.hItem);
+			}
 		}
 	} else {
-		shrink = true;
-		int extra = Math.max (4, (count + 3) / 4 * 4);
-		TreeItem [] newItems = new TreeItem [items.length + extra];
-		System.arraycopy (items, 0, newItems, 0, items.length);
-		items = newItems;
-		for (int i=itemCount; i<count; i++) {
-			new TreeItem (this, SWT.NONE, hParent, OS.TVI_LAST, 0);
+		// For performance reasons, reserve the necessary space in items[]
+		int freeCapacity = itemsGetFreeCapacity();
+		if (numInserted > freeCapacity)
+			itemsGrowArray (items.length + numInserted - freeCapacity);
+
+		// Note: on Windows, insert complexity is O(pos), so for performance
+		// reasons, all items are inserted at minimum possible position, that
+		// is, all at the same position.
+		if ((style & SWT.VIRTUAL) != 0) {
+			for (int i = 0; i < numInserted; i++) {
+				// By inserting item in the middle, the relation of
+				// 'lastIndexOf' to 'hLastIndexOf' becomes invalid if
+				// 'hLastIndexOf' points after insertion point (because
+				// inserting means that all subsequent indices change).
+				// The solution is to adjust variables.
+				if (itemFirstChild != 0) {
+					hFirstIndexOf = itemFirstChild;
+					hLastIndexOf  = itemInsertAfter;
+					lastIndexOf   = indexInsertAfter;
+				} else {
+					// There are no items whose indices are unaffected by
+					// inserting. For simplicity, reset cached index values.
+					hFirstIndexOf = 0;
+					hLastIndexOf  = 0;
+					lastIndexOf   = 0;
+				}
+
+				/*
+				 * Bug 206806: Windows sends 'TVN_GETDISPINFO' when item is
+				 * being inserted. This causes 'SWT.SetData' to be sent to
+				 * user code, but user code will likely be confused by
+				 * inconsistent Tree state (because we're still inserting):
+				 * - 'getItemCount()' will be wrong
+				 * - 'Event.index' will be wrong
+				 * The workaround is to temporarily suppress 'SWT.SetData'. Note
+				 * that the boolean flag is misleadingly used for multiple
+				 * purposes. What really happens is that 'TVN_GETDISPINFO' will
+				 * queue a repaint for item and early return.
+				 */
+				if (expanded) ignoreShrink = true;
+				createItem (null, hParent, itemInsertAfter, 0);
+				if (expanded) ignoreShrink = false;
+			}
+		} else {
+			for (int i = 0; i < numInserted; i++) {
+				new TreeItem (this, SWT.NONE, hParent, itemInsertAfter, 0);
+			}
 		}
 	}
 	if (redraw) {