… tests/org.eclipse.swt.tests/ManualTests/org/eclipse/swt/tests/manual/Issue0287_Tree_indexOf_Negative.java