@niraj-modi 2022-08-29
@Phillipus @vogella 
Let's test this patch: specifically self-hosted Eclipse use-cases.
Anyways we have some time, targeted it for 4.26 M1.