@akurtakov 2022-09-12
Please do . I don't have access to Windows to test so can't verify it.