@iloveeclipse 2022-09-19
> OK, merged it. Hopefully it will be a success this time

Sorry for that, I fear we have regression, see #396